// src/models/fraud-assessment/RiskFactorType.ts
export enum RiskFactorType {
  PHONE_INVALID = 'PHONE_INVALID',
  PHONE_DISCONNECTED = 'PHONE_DISCONNECTED',
  PHONE_RESIDENT_MISMATCH = 'PHONE_RESIDENT_MISMATCH',
  PHONE_ADDRESS_MISMATCH = 'PHONE_ADDRESS_MISMATCH',
  PHONE_SUBSCRIBER_DECEASED = 'PHONE_SUBSCRIBER_DECEASED',
  PHONE_PREPAID = 'PHONE_PREPAID',
  PHONE_NONFIXED_VOIP = 'PHONE_NONFIXED_VOIP',
  ADDRESS_INVALID = 'ADDRESS_INVALID',
  ADDRESS_INACTIVE = 'ADDRESS_INACTIVE',
  ADDRESS_RESIDENT_MISMATCH = 'ADDRESS_RESIDENT_MISMATCH',
  ADDRESS_RESIDENT_DECEASED = 'ADDRESS_RESIDENT_DECEASED',
  ADDRESS_FORWARDER = 'ADDRESS_FORWARDER',
  EMAIL_INVALID = 'EMAIL_INVALID',
  EMAIL_AUTO_GENERATED = 'EMAIL_AUTO_GENERATED',
  EMAIL_DISPOSABLE = 'EMAIL_DISPOSABLE',
  EMAIL_NAME_MISMATCH = 'EMAIL_NAME_MISMATCH',
  IP_INVALID = 'IP_INVALID',
  IP_PROXY = 'IP_PROXY',

  NONE = 'NONE',
  COUNTRY_HIGH_RISK = 'COUNTRY_HIGH_RISK',
  COUNTRY_IIN_MATCH = 'COUNTRY_IIN_MATCH',
  PRE_PAID_CARD = 'PRE_PAID_CARD',
  DEVICE_UNIQUENESS = 'DEVICE_UNIQUENESS',
  EMAIL_FREE  = 'EMAIL_FREE',
  EMAIL_HIGH_RISK = 'EMAIL_HIGH_RISK',
  ADDRESS_HIGH_RISK = 'ADDRESS_HIGH_RISK',
  ADDRESS_CITY_MISMATCH = 'ADDRESS_CITY_MISMATCH',
  ADDRESS_IP_PROXIMITY = 'ADDRESS_IP_PROXIMITY',
  ADDRESS_SHIPPING_PROXIMITY = 'ADDRESS_SHIPPING_PROXIMITY',
  ADDRESS_IP_COUNTRY_MISMATCH = 'ADDRESS_IP_COUNTRY_MISMATCH',
  MEDIUM_RISK_TRANSACTION = 'MEDIUM_RISK_TRANSACTION',
  HIGH_RISK_TRANSACTION = 'HIGH_RISK_TRANSACTION',
  EXTREME_RISK_TRANSACTION = 'EXTREME_RISK_TRANSACTION',
  PUBLIC_PROXY = 'PUBLIC_PROXY',
  ANONYMOUS_PROXY = 'ANONYMOUS_PROXY',
  IP_HIGH_RISK = 'IP_HIGH_RISK',

  IP_BLOCK_LIST = 'IP_BLOCK_LIST',
  IP_COLOCATION = 'IP_COLOCATION',
  IP_TOR = 'IP_TOR',
  IP_UNIVERSITY = 'IP_UNIVERSITY',
  PUBLISHER_COLLUSION = 'PUBLISHER_COLLUSION',
  FAST_CLICK = 'FAST_CLICK',
  REPEAT_CLICK = 'REPEAT_CLICK',
  UA_FREQUENCY = 'UA_FREQUENCY',
  IP_USER_ID_ROTATION = 'IP_USER_ID_ROTATION',
  UA_ROTATION = 'UA_ROTATION',
  AD_PROVIDER_BOT = 'AD_PROVIDER_BOT',
  PUBLISHER_REPUTATION = 'PUBLISHER_REPUTATION',
  KEYWORD_COLLUSION = 'KEYWORD_COLLUSION',
  UA_KNOWN_BOT = 'UA_KNOWN_BOT',
  UA_SPOOFED = 'UA_SPOOFED',
  VISIBILITY = 'VISIBILITY',
  COOKIE_ROTATION = 'COOKIE_ROTATION',
  CLICK_COUNT = 'CLICK_COUNT',
  SEARCH_ENGINE_INVALID = 'SEARCH_ENGINE_INVALID',
  HIDDEN = 'HIDDEN',
  REFERRER_BLACKLISTED = 'REFERRER_BLACKLISTED',
  IP_KNOWN_MALWARE = 'IP_KNOWN_MALWARE',
}